feat: centralize env file loader
