+++ alpha_factory_v1/utils/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Utility helpers for Alpha-Factory."""
+
+__all__ = ["env"]
