+++ src/utils/config.py
@@ -5,34 +5,14 @@ from __future__ import annotations
 import logging
 import os
 from pathlib import Path
-from typing import Any, Dict, Optional, cast
+from typing import Any, Optional, cast
+
+from alpha_factory_v1.utils.env import _load_env_file
 
 from pydantic import Field
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 _log = logging.getLogger(__name__)
-
-
-def _load_env_file(path: str | os.PathLike[str]) -> Dict[str, str]:
-    """Return key/value pairs from ``path``.
-
-    Falls back to a minimal parser when :mod:`python_dotenv` is unavailable.
-    """
-    try:  # pragma: no cover - optional dependency
-        from dotenv import dotenv_values
-
-        return {k: v for k, v in dotenv_values(path).items() if v is not None}
-    except Exception:  # noqa: BLE001 - any import/parsing error falls back
-        pass
-
-    data: Dict[str, str] = {}
-    for line in Path(path).read_text(encoding="utf-8").splitlines():
-        line = line.strip()
-        if not line or line.startswith("#") or "=" not in line:
-            continue
-        k, v = line.split("=", 1)
-        data[k.strip()] = v.strip().strip('"')
-    return data
 
 
 def _load_dotenv(path: str = ".env") -> None:
